requirements.txt
cbor2==5.2.0
digi-xbee==1.3.0
pika==1.1.0
pycryptodome==3.9.9
pyserial==3.5
requests==2.25.1
